[MCH] improve logic of slot finalization
The check of the slot statistics and the slot finalization are explicitly
done in the calibrator code instead of relying on the `checkSlotsToFinalize()`
function, because for MCH we only have one slot that spans the whole run.

`setUpdateAtTheEndOfRunOnly()` is also called at initialization to skip the
internal check of the slot statistics at each TF.

In the case where the calibrations are produced as soon as a target statistics
is reached, the subsequent data is skipped to avoid producing multiple
calibration objects for the same run.