[MCH] use mapping in pedestals calibration
The electronics mapping is used to flag as valid the readout channels that
are associated to detector pads. Only valid channels are processed and
used to compute the total pedestals data size. Invalid channels therefore
will not contribute to the computation when checking the statistics of
a given calibration slot.