[MCH] added option to output the time-clustered ROFs without filtering
The additional output containing all the time-clustered ROFs without futher filtering can be useful for diagnistics, fir example in online QC.
The output is disabled by default, and can be enabled by setting `"MCHTimeClusterizer.enableUnfilteredOutput=true"` in the configuration keys.
The default name of the DPL message is `"UTC-F-DIGITROFS"`. It can be customized via the `--output-unfiltered-digitrofs-data-description` command-line option.